Dev/create user management presentation layer

+++ src/main/kotlin/api/masterplan/app/userManagementModule/application/service/UserServiceImpl.kt
@@ -1,6 +1,8 @@
 package api.masterplan.app.userManagementModule.application.service
 
 import api.masterplan.app.logging.LoggingMethod
+import api.masterplan.app.userManagementModule.application.dto.EmployeeInfo
+import api.masterplan.app.userManagementModule.application.ports.EmployeeCreationPort
 import api.masterplan.app.userManagementModule.domain.dtos.AppUserDetails
 import api.masterplan.app.userManagementModule.domain.exceprions.UserManagementException
 import api.masterplan.app.userManagementModule.domain.interfaces.UserRepository
@@ -15,39 +17,55 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 class UserServiceImpl(
-    private val userRepository: UserRepository
+    private val userRepository: UserRepository,
+    private val employeeCreationPort: EmployeeCreationPort
 ): UserService {
 
-    @LoggingMethod
-    @Transactional(rollbackFor = [Exception::class])
-    override fun getUserByLogin(login: UserLogin): AppUserDetails {
-        val user = userRepository.findByLogin(login) ?: throw UserManagementException.UserNotFoundException(login)
+    @LoggingMethod(moduleName = "userManagementModule")
+    override fun getUserById(userId: UserId): AppUserDetails {
+        val user = userRepository.getUser(userId) ?: throw UserManagementException.UserNotExistsException(userId)
         val appUserDetails = AppUserDetails(
             id = user.id,
-            login = login,
+            login = user.login,
             password = user.password,
             roles = user.roles
         )
         return appUserDetails
     }
 
-    @LoggingMethod
+    @LoggingMethod(moduleName = "userManagementModule")
+    @Transactional(rollbackFor = [Exception::class])
+    override fun getUserByLogin(login: UserLogin): AppUserDetails {
+        val user = userRepository.findByLogin(login) ?: throw UserManagementException.UserNotFoundException(login)
+        val appUserDetails = AppUserDetails(
+            id = user.id,
+            login = user.login,
+            password = user.password,
+            roles = user.roles
+        )
+        return appUserDetails
+    }
+
+    @LoggingMethod(moduleName = "userManagementModule")
     @Transactional(rollbackFor = [Exception::class])
     override fun resetPasswordForUser(userId: UserId, newPassword: UserPassword): UserId {
+        userRepository.getUser(userId)?: throw UserManagementException.UserNotExistsException(userId)
         val editedUserId = userRepository.setPassword(userId,newPassword) ?: throw UserManagementException.FailedToResetPasswordForUser(userId)
         return editedUserId
     }
 
-    @LoggingMethod
+    @LoggingMethod(moduleName = "userManagementModule")
     @Transactional(rollbackFor = [Exception::class])
-    override fun createUser(login: UserLogin, password: UserPassword, roles: Set<UserRole>): UserId {
+    override fun createUser(login: UserLogin, password: UserPassword,
+                            roles: Set<UserRole>,employeeInfo: EmployeeInfo): UserId {
         if (userRepository.isUserExist(login)) throw UserManagementException.UserAlreadyExistsException(login)
         val newUser = AppUser.create(login = login, rawPassword = password, roles = roles)
         val userId = userRepository.saveUser(newUser) ?: throw UserManagementException.FailedToCreateUserException(login)
+        employeeCreationPort.createEmployee(userId,employeeInfo)
         return userId
     }
 
-    @LoggingMethod
+    @LoggingMethod(moduleName = "userManagementModule")
     override fun getUser(userId: UserId): AppUserDetails {
         val user = userRepository.getUser(userId) ?: throw UserManagementException.UserNotExistsException(userId)
         val userDetails = AppUserDetails(
@@ -59,9 +77,11 @@ class UserServiceImpl(
         return userDetails
     }
 
-    @LoggingMethod
+    @LoggingMethod(moduleName = "userManagementModule")
     @Transactional(rollbackFor = [Exception::class])
     override fun deleteUser(userId: UserId): UserId {
+        val user = userRepository.getUser(userId)?: throw UserManagementException.UserNotExistsException(userId)
+        if (user.isAdmin()) throw UserManagementException.UserCantBeDeleted(userId)
         val deletedUserId = userRepository.deleteUser(userId) ?: throw UserManagementException.FailedToDeleteUserException(userId)
         return deletedUserId
     }
